game_entity: reduce size of stored image indexes to uint8_t
Size reduction on AVR: 330 -> 242 bytes

+++ src/game_entity.h
@@ -77,7 +77,7 @@ namespace GameEntities {
 
         // All image frames used to draw the entity.  Each image is represented
         // by a unique integer handle.
-        uint16_t images[NUM_GAME_ENTITY_IMAGES];
+        uint8_t images[NUM_GAME_ENTITY_IMAGES];
     };
 
     class GameEntity {
